\encoding{UTF-8}
\name{DetSel-internal}
\alias{cumulative.distribution.of.probabilities}
\alias{filled.contour3}
\alias{GetData}
\alias{get.simulation.parameters}
\alias{make.2D.histogram}
\alias{draw.single.detsel.graph}
\alias{SimulDiv}
\alias{write.detsel.file}
\title{Internal DetSel Function}
\description{
This function is not intended to be called by the user.
}


