\name{make.example.files}
\Rdversion{1.0}
\alias{make.example.files}
\title{Generate an Example File in the Working Directory}
\description{This command will copy two example files into the user's working directory. The file "data.dat" is in DetSel format, and the file "data.gen" is in Genepop format (see Raymond and Rousset 1995; Rousset 2007).
}
\usage{
  make.example.files()
}
\value{
The example file is saved in the current directory.
}
\examples{
## This is to generate an example file in the working directory.
make.example.files()
}
\references{
Raymond, M., and Rousset, F. (1995) Genepop (version 1.2): population genetics software for exact tests and ecumenicism, \emph{Journal of Heredity} \bold{86}: 248-249.

Rousset, F. (2008) genepop'007: a complete re-implementation of the genepop software for Windows and Linux, \emph{Molecular Ecology Notes} \bold{8}: 103-106.
}