\name{remove.example.files}
\Rdversion{1.0.4}
\alias{remove.example.files}
\title{Remove Example Files from the Working Directory}
\description{This command will remove example files from the working directory.
}
\usage{
  remove.example.files()
}
\details{
The example files generated by \code{\link{make.example.files}} and further analyses are removed from the current directory.
}
\seealso{
\code{\link{make.example.files}}
}
